separate zoom functionary to ZoomableImage component
adjust styling of ImageLoader
add styling for ZoomableImage
